+++ superset/assets/src/visualizations/Table/ReactTable.js
@@ -1,0 +1,4 @@
+import reactify from '../../utils/reactify';
+import Component from './Table';
+
+export default reactify(Component);
