+++ superset/assets/src/visualizations/Table/Table.css
