+++ superset/assets/src/visualizations/Table/adaptor.jsx
@@ -1,0 +1,5 @@
+import createAdaptor from '../../utils/createAdaptor';
+import Component from './ReactTable';
+import transformProps from './transformProps';
+
+export default createAdaptor(Component, transformProps);
