+++ superset/assets/src/visualizations/Table/transformProps.js
@@ -1,0 +1,58 @@
+export default function transformProps(basicChartInput) {
+  const {
+    datasource,
+    filters,
+    formData,
+    onAddFilter,
+    payload,
+  } = basicChartInput;
+  const {
+    alignPn,
+    colorPn,
+    includeSearch,
+    metrics,
+    orderDesc,
+    pageLength,
+    percentMetrics,
+    tableFilter,
+    tableTimestampFormat,
+    timeseriesLimitMetric,
+  } = formData;
+  const { columnFormats, verboseMap } = datasource;
+  const { records, columns } = payload.data;
+
+  const processedColumns = columns.map((key) => {
+    let label = verboseMap[key];
+    // Handle verbose names for percents
+    if (!label) {
+      if (key[0] === '%') {
+        const cleanedKey = key.substring(1);
+        label = '% ' + (verboseMap[cleanedKey] || cleanedKey);
+      } else {
+        label = key;
+      }
+    }
+    return {
+      key,
+      label,
+      format: columnFormats && columnFormats[key],
+    };
+  });
+
+  return {
+    data: records,
+    alignPositiveNegative: alignPn,
+    colorPositiveNegative: colorPn,
+    columns: processedColumns,
+    filters,
+    includeSearch,
+    metrics,
+    onAddFilter,
+    orderDesc,
+    pageLength: pageLength && parseInt(pageLength, 10),
+    percentMetrics,
+    tableFilter,
+    tableTimestampFormat,
+    timeseriesLimitMetric,
+  };
+}
