+++ superset/assets/src/visualizations/models/BasicChartInput.js
@@ -4,7 +4,11 @@ export default class BasicChartInput {
   constructor(slice, payload, setControlValue) {
     this.annotationData = slice.annotationData;
     this.datasource = convertKeysToCamelCase(slice.datasource);
+    this.filters = slice.getFilters();
     this.formData = convertKeysToCamelCase(slice.formData);
+    this.onAddFilter = (...args) => {
+      slice.addFilter(...args);
+    };
     this.payload = payload;
     this.setControlValue = setControlValue;
   }
